Fix CompoundTranslogHeader.readHeaderOld (#491)
* Fix CompoundTranslogHeader.readHeaderOld

* Fix both old and new

+++ x-pack/plugin/stateless/src/test/java/co/elastic/elasticsearch/stateless/engine/CompoundTranslogHeaderTests.java
@@ -142,7 +142,7 @@ public class CompoundTranslogHeaderTests extends AbstractWireSerializingTestCase
         ) {
             // Serialize in old version
             out.writeMap(testInstance.metadata());
-            out.writeInt((int) out.getChecksum());
+            out.writeLong(out.getChecksum());
             out.flush();
 
             try (StreamInput in = output.bytes().streamInput()) {
